Refactoring code after first testing
