+++ src/App.js
@@ -8,13 +8,13 @@ import {
   createRoutesFromElements,
 } from "react-router-dom";
 import Home from "./pages/Home";
-import Cart from "./components/cart/Cart";
+import Cart from "./components/main/cart/Cart";
 import Signin from "./pages/Signin";
 import Error from "./pages/Error";
 import Registration from "./pages/Registration";
-import Header from "./components/header/Header";
-import Footer from "./components/footer/Footer";
-import Variant from "./pages/Variant";
+import Header from "./components/common/header/Header";
+import Footer from "./components/common/footer/Footer";
+import ProductDetail from "./components/main/variant/ProductDetail";
 
 const Layout = () => {
   return (
@@ -37,8 +37,8 @@ function App() {
         <Route path="/" element={<Layout />}>
           <Route index element={<Home />} />
           <Route path="/cart" element={<Cart />} />
+          <Route path="/product/:id" element={<ProductDetail />} />
         </Route>
-        <Route path="/product/:id" element={<Variant />} />
       </Fragment>
     )
   );
