+++ src/components/main/variant/ProductDetail.js
@@ -3,8 +3,8 @@ import StarIcon from "@mui/icons-material/Star";
 import FmdGoodIcon from "@mui/icons-material/FmdGood";
 import { Link, useParams } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
-import { getVariant, selectError, selectLoading, selectSuccess, selectVariantDetail } from "../../features/variant/variantSlice"
-import { addToCart } from "../../features/cart/cartSlice";
+import { getVariant, selectError, selectLoading, selectSuccess, selectVariantDetail } from "../../../features/variant/variantSlice"
+import { addToCart } from "../../../features/cart/cartSlice";
 
 function ProductDetail() {
   const {id} = useParams();
